third_party/WebKit/LayoutTests/external/wpt/notifications/idlharness.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
FAIL idl_test setup promise_test: Unhandled rejection with value: object "TypeError: Failed to construct 'Notification': Illegal constructor."
PASS Partial interface ServiceWorkerRegistration: original interface defined
PASS Partial interface ServiceWorkerGlobalScope: original interface defined
PASS Notification interface: existence and properties of interface object
PASS Notification interface object length
PASS Notification interface object name
PASS Notification interface: existence and properties of interface prototype object
PASS Notification interface: existence and properties of interface prototype object's "constructor" property
PASS Notification interface: existence and properties of interface prototype object's @@unscopables property
PASS Notification interface: attribute permission
PASS Notification interface: member requestPermission
PASS Notification interface: attribute maxActions
PASS Notification interface: attribute onclick
PASS Notification interface: attribute onshow
PASS Notification interface: attribute onerror
PASS Notification interface: attribute onclose
PASS Notification interface: attribute title
PASS Notification interface: attribute dir
PASS Notification interface: attribute lang
PASS Notification interface: attribute body
PASS Notification interface: attribute tag
PASS Notification interface: attribute image
PASS Notification interface: attribute icon
PASS Notification interface: attribute badge
PASS Notification interface: attribute vibrate
PASS Notification interface: attribute timestamp
PASS Notification interface: attribute renotify
PASS Notification interface: attribute silent
PASS Notification interface: attribute requireInteraction
PASS Notification interface: attribute data
PASS Notification interface: attribute actions
PASS Notification interface: operation close()
FAIL Notification must be primary interface of notification assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Stringification of notification assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "permission" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must not have property "requestPermission" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "maxActions" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "onclick" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "onshow" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "onerror" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "onclose" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "title" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "dir" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "lang" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "body" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "tag" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "image" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "icon" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "badge" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "vibrate" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "timestamp" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "renotify" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "silent" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "requireInteraction" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "data" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "actions" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
FAIL Notification interface: notification must inherit property "close()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notification is not defined"
PASS NotificationEvent interface: existence and properties of interface object
PASS NotificationEvent interface object length
PASS NotificationEvent interface object name
PASS NotificationEvent interface: existence and properties of interface prototype object
PASS NotificationEvent interface: existence and properties of interface prototype object's "constructor" property
PASS NotificationEvent interface: existence and properties of interface prototype object's @@unscopables property
PASS NotificationEvent interface: attribute notification
PASS NotificationEvent interface: attribute action
FAIL NotificationEvent must be primary interface of notificationEvent assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notificationEvent is not defined"
FAIL Stringification of notificationEvent assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notificationEvent is not defined"
FAIL NotificationEvent interface: notificationEvent must inherit property "notification" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notificationEvent is not defined"
FAIL NotificationEvent interface: notificationEvent must inherit property "action" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: notificationEvent is not defined"
PASS ServiceWorkerRegistration interface: operation showNotification(DOMString, NotificationOptions)
PASS ServiceWorkerRegistration interface: operation getNotifications(GetNotificationOptions)
PASS ServiceWorkerGlobalScope interface: attribute onnotificationclick
PASS ServiceWorkerGlobalScope interface: attribute onnotificationclose
PASS ServiceWorkerGlobalScope interface: self must inherit property "onnotificationclick" with the proper type
PASS ServiceWorkerGlobalScope interface: self must inherit property "onnotificationclose" with the proper type
PASS HTMLElement interface: existence and properties of interface object
PASS HTMLBodyElement interface: existence and properties of interface object
PASS Window interface: existence and properties of interface object
PASS HTMLFrameSetElement interface: existence and properties of interface object
PASS Node interface: existence and properties of interface object
PASS Document interface: existence and properties of interface object
PASS DocumentType interface: existence and properties of interface object
PASS DocumentFragment interface: existence and properties of interface object
PASS ShadowRoot interface: existence and properties of interface object
PASS Element interface: existence and properties of interface object
PASS CharacterData interface: existence and properties of interface object
PASS Text interface: existence and properties of interface object
Harness: the test ran to completion.
